Detach from dop251 master branch
This updates all imports to refer to the fork at github.com/nilium/goja
and adds a notice in the README regarding this.

goja_nodejs is removed as a dependency from goja/goja.

.bak has also been added to the .gitignore, just because that was used
when changing the import paths via sed.

This branch is intended to be used in the event that the goja author
refuses to consider fixing the data race around the interrupt check in
the VM's runloop.