tests: Run tc39 tests as parallel subtests
This removes the program cache -- it's not worth it and means tests
can't run in parallel. Subtests are handled via the runSubtest
function. Parallel tests are handled by check if the TB implements
parallelizer under setParallelTest.

Most of this is just to account for TB being used everywhere, but may
also help a little bit with early testing package versions.